+++ src/Components/Apply/Apply.js
@@ -4,7 +4,7 @@ import Check from "./Check";
 import { equipment_table } from "../../Data/equipment_table.json";
 import { useHistory } from "react-router";
 import ModalPortal from "../ModalPortal/ModalPortal";
-import useModal from "../../Hooks/useModal";
+import useModal from "../../Hooks/UseModal";
 
 const oneThings = ["노트북", "태블릿", "모니터"];
